requirements.txt
psutil==5.9.8
py-cpuinfo==9.0.0
wmi==1.5.1; platform_system == "Windows"
pywin32==306; platform_system == "Windows"
